[#389] Extract SimulationTime parity helpers into tests/common/
Move `assert_simulation_time_bits_eq` and `bevy_sim_time` from
`bevy_parity_timescale.rs` and `bevy_parity_time_docker.rs` into the
shared `tests/common/mod.rs` so the 13-field bit-identity check has
one definition. Both wrappers had byte-for-byte identical copies; the
shared helper preserves the existing runner/bevy panic-message format
minus the file-name prefix (cargo nextest already identifies the
binary in failure output).
